// frontend/src/app/api/model/createFeatureInput.ts
/**
 * Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateFeatureInput { 
    title: string;
    description?: string | null;
    epicLink?: string | null;
    priority: number;
    deadline?: string | null;
}